Update data_reader.py so you can pass smell as argument
There was a problem that when I ran collector with ssh and i either closed the terminal or my pc went to sleep it turned off the process. 
To run process without it shutting down and not downloading anything else I can pass it as argument and use linux nohup command to make it run.